Split SignsHelper into smaller helper modules
